refactor: enhance folder structure
