+++ bin/deploy.ts
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 import 'source-map-support/register';
 import * as cdk from 'aws-cdk-lib';
-import { AppStack } from '../lib/app-stack';
+import { AppStack } from '../src/stack.default';
 
 const app = new cdk.App();
 new AppStack(app, 'RSS-Back', {
