+++ cdk.json
@@ -1,5 +1,5 @@
 {
-  "app": "npx ts-node --prefer-ts-exts bin/app.ts",
+  "app": "npx ts-node --prefer-ts-exts bin/deploy.ts",
   "watch": {
     "include": [
       "**"
