+++ test/src/lambda/products/get-list.test.ts
@@ -1,5 +1,5 @@
 import { getProductsList } from "~/lambda/products/get-list";
-import productService from "~/services/repositories/product";
+import productService from "~/services/db/product";
 
 import { baseEvent, baseContext } from "./utils";
