+++ test/src/lambda/products/utils.ts
