+++ tsconfig.json
@@ -24,7 +24,7 @@
       "./node_modules/@types"
     ],
     "paths": {
-      "~/*": ["./*"]
+      "~/*": ["./src/*"]
     },
   },
   "exclude": [
